Fix tokio-openssl connect API
After creating the SslStream it first has to be connected before it can
actually be used to initiate the WebSocket connection.

While this technically is a breaking API changes because of wrapping the
stream in a Pin<Box<_>> and having this as part of the public API, the
old API simply didn't work so if anybody's code fails to compile now
they would've had non-functioning code anyway.